Add initial analyzer

+++ src/ChaosDbg.Tests/Services/Disasm/BaseDisasmTest.cs
@@ -14,7 +14,7 @@ namespace ChaosDbg.Tests
             {
                 OptionalHeader = new MockImageOptionalHeader
                 {
-                    ImageBase = (ulong) ip
+                    ImageBase = ip
                 }
             };
